# Sandbox: Avoid redundant extra dependencies