Closes N/A

<!-- If your PR is related to an issue, provide the number(s) above; if it resolves multiple issues, be sure to break them up (e.g. "closes #1000, closes #1001"). -->

<!--

Thank you for contributing to Storybook! Please submit all PRs to the `next` branch unless they are specific to the current release. Storybook maintainers cherry-pick bug and documentation fixes into the `main` branch as part of the release process, so you shouldn't need to worry about this. For additional guidance: https://storybook.js.org/docs/contribute

-->

## What I did

Avoid redundant extra dependencies

## Checklist for Contributors

### Testing

<!-- Please check (put an "x" inside the "[ ]") the applicable items below to communicate how to test your changes -->

#### The changes in this PR are covered in the following automated tests:
- [ ] stories
- [ ] unit tests
- [ ] integration tests
- [ ] end-to-end tests

#### Manual testing

_This section is mandatory for all contributions. If you believe no manual test is necessary, please state so explicitly. Thanks!_

<!-- Please include the steps to test your changes here. For example:

1. Run a sandbox for template, e.g. `yarn task --task sandbox --start-from auto --template react-vite/default-ts`
2. Open Storybook in your browser
3. Access X story

-->

### Documentation

<!-- Please check (put an "x" inside the "[ ]") the applicable items below to indicate which documentation has been updated. -->

- [ ] Add or update documentation reflecting your changes
- [ ] If you are deprecating/removing a feature, make sure to update
      [MIGRATION.MD](https://github.com/storybookjs/storybook/blob/next/MIGRATION.md)

## Checklist for Maintainers

- [ ] When this PR is ready for testing, make sure to add `ci:normal`, `ci:merged` or `ci:daily` GH label to it to run a specific set of sandboxes. The particular set of sandboxes can be found in `code/lib/cli/src/sandbox-templates.ts`
- [ ] Make sure this PR contains **one** of the labels below:
   <details>
     <summary>Available labels</summary>

     - `bug`: Internal changes that fixes incorrect behavior.
     - `maintenance`: User-facing maintenance tasks.
     - `dependencies`: Upgrading (sometimes downgrading) dependencies.
     - `build`: Internal-facing build tooling & test updates. Will not show up in release changelog.
     - `cleanup`: Minor cleanup style change. Will not show up in release changelog.
     - `documentation`: Documentation **only** changes. Will not show up in release changelog.
     - `feature request`: Introducing a new feature.
     - `BREAKING CHANGE`: Changes that break compatibility in some way with current major version.
     - `other`: Changes that don't fit in the above categories.
   
   </details>

### 🦋 Canary release

<!-- CANARY_RELEASE_SECTION -->

This PR does not have a canary release associated. You can request a canary release of this pull request by mentioning the `@storybookjs/core` team here.

_core team members can create a canary release [here](https://github.com/storybookjs/storybook/actions/workflows/canary-release-pr.yml) or locally with `gh workflow run --repo storybookjs/storybook canary-release-pr.yml --field pr=<PR_NUMBER>`_

<!-- CANARY_RELEASE_SECTION -->

<!-- BENCHMARK_SECTION -->

| name | before | after | diff | z | % |
| ---- | ------ | ----- | ---- | - | - |
| createSize |  0 B | 0 B | 0 B | - | - |
| generateSize |  77.3 MB | 77.3 MB | 0 B | 0.49 | 0% |
| initSize |  162 MB | 162 MB | 0 B | 0.64 | 0% |
| diffSize |  85 MB | 85 MB | 0 B | 0.68 | 0% |
| buildSize |  7.58 MB | 7.58 MB | 0 B | **1.5** | 0% |
| buildSbAddonsSize |  1.67 MB | 1.67 MB | 0 B | **1.53** | 0% |
| buildSbCommonSize |  195 kB | 195 kB | 0 B | - | 0% |
| buildSbManagerSize |  2.34 MB | 2.34 MB | 0 B | 1.22 | 0% |
| buildSbPreviewSize |  352 kB | 352 kB | 0 B | 1 | 0% |
| buildStaticSize |  0 B | 0 B | 0 B | - | - |
| buildPrebuildSize |  4.55 MB | 4.55 MB | 0 B | **1.53** | 0% |
| buildPreviewSize |  3.02 MB | 3.02 MB | 0 B | 1 | 0% |
| testBuildSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbAddonsSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbCommonSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbManagerSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbPreviewSize |  0 B | 0 B | 0 B | - | - |
| testBuildStaticSize |  0 B | 0 B | 0 B | - | - |
| testBuildPrebuildSize |  0 B | 0 B | 0 B | - | - |
| testBuildPreviewSize |  0 B | 0 B | 0 B | - | - |



| name | before | after | diff | z | % |
| ---- | ------ | ----- | ---- | - | - |
| createTime |  6.8s | 22s | 15.1s | **1.55** | 🔺68.8% |
| generateTime |  19.7s | 22.4s | 2.6s | 0.92 | 11.8% |
| initTime |  16.8s | 18s | 1.2s | 0.49 | 6.9% |
| buildTime |  11.9s | 11.5s | -396ms | -0.1 | -3.4% |
| testBuildTime |  0ms | 0ms | 0ms | - | - |
| devPreviewResponsive |  6.1s | 8.5s | 2.4s | **1.84** | 🔺28.2% |
| devManagerResponsive |  3.9s | 5.4s | 1.4s | **1.71** | 🔺27.5% |
| devManagerHeaderVisible |  693ms | 988ms | 295ms | 1.12 | 29.9% |
| devManagerIndexVisible |  725ms | 989ms | 264ms | 0.83 | 26.7% |
| devStoryVisibleUncached |  1s | 1.1s | 52ms | -0.77 | 4.6% |
| devStoryVisible |  724ms | 1s | 302ms | 1.07 | 29.4% |
| devAutodocsVisible |  621ms | 843ms | 222ms | 1.18 | 26.3% |
| devMDXVisible |  621ms | 827ms | 206ms | 0.94 | 24.9% |
| buildManagerHeaderVisible |  736ms | 832ms | 96ms | 0.73 | 11.5% |
| buildManagerIndexVisible |  767ms | 1s | 289ms | **2.53** | 🔺27.4% |
| buildStoryVisible |  768ms | 1s | 290ms | **2.56** | 🔺27.4% |
| buildAutodocsVisible |  603ms | 886ms | 283ms | **1.73** | 🔺31.9% |
| buildMDXVisible |  594ms | 911ms | 317ms | **3.34** | 🔺34.8% |

<!-- BENCHMARK_SECTION -->

<!-- greptile_comment -->

## Greptile Summary

This PR optimizes the handling of extra dependencies in the sandbox creation process by introducing a Set to store unique dependencies, avoiding duplicates.

- Introduced `extraDeps` Set in `scripts/tasks/sandbox.ts` to prevent redundant dependencies
- Used Set's `add` method for individual additions and spread operator for initial population
- Converted Set back to array before passing to `addExtraDependencies` function
- Improved efficiency by eliminating duplicate entries in extra dependencies list
- Simplified code structure for better readability and maintainability

<!-- /greptile_comment -->